fix: transaction atomicity and integer overflow in billing
- Fix integer overflow in stats view: cast time_total to bigint before
  multiplication to prevent overflow with high-volume aggregated logs
  (timescale-setup.ts)

- Fix transaction atomicity in addUsageChargesToDraft: change parameter
  to DatabaseOrTransaction requiring caller to provide transaction context,
  preventing double-billing on partial failures (usage-charges.ts)

- Update getBillableRequestCount to accept DatabaseOrTransaction for
  consistency with billing transaction requirements (queries.ts)

- Fix transaction atomicity in subscribe endpoint: wrap post-billing
  writes (subscriptionChargePending update + ledger entry) in transaction
  to prevent double-billing if server crashes after payment (services.ts)

+++ packages/database/src/timescale-setup.ts
@@ -74,7 +74,7 @@ export async function setupTimescaleDB() {
       -- Legacy/summary (all 2xx)
       SUM(repeat) FILTER (WHERE status_code >= 200 AND status_code < 300) AS success_count,
       -- Performance (weighted average for pre-aggregated logs)
-      SUM(time_total * repeat)::double precision / NULLIF(SUM(repeat), 0) AS avg_response_time_ms,
+      SUM(time_total::bigint * repeat)::double precision / NULLIF(SUM(repeat), 0) AS avg_response_time_ms,
       SUM(bytes_sent * repeat) AS total_bytes
     FROM haproxy_raw_logs
     WHERE customer_id IS NOT NULL
